Add accessibility statement
The accessibility statement is required to be placed in a proeminent
position on the website and that it can be accessed from all pages.

The accessibility team's recommendation is to put it in the footer.

Using the footer links feature to add a link to the Accessibility
statement. Also hiding a second link to the accessibility statement in
the left hand navigation so it doesn't get mixed in with the technical
content.