html-proofer should ignore empty alt tags
When writing documentation, we always try to have the information in the
image also present in the surrounding text. If this is not possible,
then we write meaningful alt text.

However, in most cases images in documentation are described in
surrounding text so most of the time we need emty alt tags and we don't
want the linter to bring this up as an error.

Relevant W3C guidance:
https://www.w3.org/WAI/tutorials/images/decorative/