uSnau9tK: setup test branch for prototyping
Currently, it's just a mirror of mastre but I'll soon submit a PR to change it to do what the README now says it does.

This will allow us to make changes that we can validate with user testing before updating the 'live' version of the docs.

The username/passwaord are not for security - they're just to make it less likely that a pilot participant mistakes the prototype for the real thing.