BAU: remove references to 'clientId' from the docs
We do not use this field for anything - we just record it in our logs. This field is optional. Thus, we can omit it from the pilot docs and save pilot participants the effort of providing the field, and us the effort of needing to provide them with the value.