BAU: add support for `RSA-OAEP-256` encryption
This is the same as `RSA-OAEP`, except it uses SHA256 instead of SHA1. Both algorithms are still secure - despite SHA1 being now considered weak, it is only used as an entropy source, so doesn't need to be collision-resistant.

It is still good to officially document support for both, though. Some libraries mark `RSA-OAEP` as deprecated, but support for `RSA-OAEP-256` is not yet ubiquitous. The DCS already supports both, so there's no extra work for us to do.

In the long term, we might want to consider switching from RSA-based algorithms to elliptic-curve-based algorithms. This is because [RSA is generally considered harder to implement correctly](https://blog.trailofbits.com/2019/07/08/fuck-rsa/). However, we're not going to do that now, since it would require substantial changes to the DCS and our certificate processes that we have not yet the time for.