lK5rICbK: Add prerequisites for integration
Previously we didn't have a clear definition of what was required before a participant started to create their clients.

This commit adds such a list, with focus on outlining what certificates and keys are required before starting their onboarding process. It also clearly defines which keypairs are provided (by us) and which ones they will have to manually create themselves (or submit CSRs).

I've also added a section to the support section to help define what we expect to see from a pilot participant when they are asking for help.

Revert "Revert "lK5rICbK: Add prerequisites for integration""