xpCNF3HW: make cert links active or remove them
It's better for links to be active, since this makes them easier to understand when using a screen reader. Convert certificate links to be active. There were a couple of links where this would be more difficult, so I removed the links since the download instructions were already in the prerequisites section.

The links on the 'How the DCS works' page will be handled separately.

As recommended by Pippa.