xpCNF3HW: move instructions to download DCS certs
They were previously in a place that was a bit awkward since everything else in the list was a link to a page of instructions. It was also difficult to make the links active.

Instead, move the suggestion to download the DCS public certs into the 'generate certs' page. It's a bit questionable as to whether it belongs there, but it seems not unreasonable as another certificate-related prerequisite for all further steps.