H3WCW3iA: provide more detail on rate limits
Since the rate limit text was first written, we have agreed our initial rate limit. Since we now know how long the window will be, we can tell clients to help them prepare.

We cannot give a precise number at this stage. Commercial confidentiality means we will be unable to until all possible pilot participants have decided one way or the other. In the interim, use an indicative number that is within an order of magnitude of the expected final number (i.e. in the range 1 to 100).