FjWF1vJ0: give Ben and Beth approval rights
Make everyone in the verify-dcs-docs team a code owner. Ben and Beth (?) are the only people who will be affected by this. They will now have the power to approve PRs for this repo. They still won't be able to merge PRs, though.

Membership of the verify-dcs-docs team is managed in terraform.