BAU: add custom lint for broken link partial use
There are two ways of adding a link:

1. directly - `[foo](bar.com)`
2. using the links partial - `[bar]` or `[foo][bar]`, where the link partial has an entry for `[bar]` that points to `bar.com`.

We've had problems with broken links. This is solved for case 1, since html-proofer catches those. However, it is not able to catch broken links created by case two. These manifest as seeing `[foo][bar]` still present in the HTML output.

So create a new html-proofer checker that checks for broken links that try (but fail) to use the link partial.

This is a copy of the linter that we already using successfully in our internal docs - where it successfully caught a few broken links.