BAU: Remove confusing link and references to JOSE
The document contained a link labeled "a JOSE object" that pointed to documentation about how to create a certificate thumbprint. This is confusing. Furthermore we rarely mention JOSE and can avoid doing so.

Remove the link and all references to JOSE in our text. Technical uses in Content-Type and names of libraries remain, without comment.

Rename "Set up your JOSE certificate thumbprints" to "Generate certificate thumbprints" adn edit it to make it more generic.